Change `hab.meta.NotSet` to `hab.NotSet` and remove python 2 metaclass support
