+++ hab/parsers/__init__.py
@@ -5,7 +5,7 @@ from .distro import Distro
 from .distro_version import DistroVersion
 from .flat_config import FlatConfig
 from .hab_base import HabBase
-from .meta import HabMeta, NotSet, hab_property
+from .meta import HabMeta, hab_property
 from .placeholder import Placeholder
 
 __all__ = [
@@ -16,6 +16,5 @@ __all__ = [
     'hab_property',
     'HabBase',
     'HabMeta',
-    'NotSet',
     'Placeholder',
 ]
