+++ hab/parsers/flat_config.py
@@ -1,7 +1,8 @@
 import logging
 
+from .. import NotSet
 from .config import Config
-from .meta import NotSet, hab_property
+from .meta import hab_property
 
 logger = logging.getLogger(__name__)
